Remove: Popup and replace it to Modal
